docker-update: Improves error handling

+++ docker-update.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-set -ex
+set -x
 
 ########## Print GNU GPLv3 LICENSE notices ##########
 echo '
@@ -46,7 +46,7 @@ fi
 ############### Assign Variables ##############
 DOCKER_STACK="Docker"
 DOCKER_STATUS_URL="the Status Dashboard"
-[ -n "$COMPOSE_PROJECT_NAME" ] && DOCKER_STACK="$COMPOSE_PROJECT_NAME Docker"
+[ -n "$COMPOSE_PROJECT_NAME" ] && DOCKER_STACK="Docker $COMPOSE_PROJECT_NAME"
 [[ -n "$DASHBOARD_HOST" && -n "$DOMAINNAME" ]] && DOCKER_STATUS_URL="[Status Dashboard](https://$DASHBOARD_HOST.$DOMAINNAME)"
 [ -n "$TELEGRAM_BOT_SERVICEALERTS_TOKEN" ] && BOT_TOKEN="$TELEGRAM_BOT_SERVICEALERTS_TOKEN"
 [ -n "$TELEGRAM_BOT_SERVICEALERTS_CHATID" ] && CHAT_ID="$TELEGRAM_BOT_SERVICEALERTS_CHATID"
@@ -64,7 +64,7 @@ MESSAGE_CLEANUP="🗑️ Removing unused *__STATS__*..."
 MESSAGE_UPDATE="📥 Updating *__DOCKERIMAGE__*"
 MESSAGE_NOUPDATE="⏩ Skipping *__DOCKERIMAGE__* (no update)"
 MESSAGE_ERRUPDATE="⚠️ Something went wrong checking *__DOCKERIMAGE__*, please investigate."
-MESSAGE_RESTARTING="🔁 Restarting *__DOCKERSERVICE__*..."
+MESSAGE_RESTARTING="🔁 Restarting *${COMPOSE_PROJECT_NAME:+$COMPOSE_PROJECT_NAME-}__DOCKERSERVICE__*..."
 MESSAGE_FINISH="✅ ALL DONE: *$DOCKER_STACK* is up-to-date!
 > Check $DOCKER_STATUS_URL"
 
@@ -84,8 +84,24 @@ send_telegram_message() {
          -F message_thread_id="$CHAT_TOPIC_ID" \
          -F parse_mode=MarkdownV2 \
          "https://api.telegram.org/bot$BOT_TOKEN/sendMessage" 2>&1 || true
+    else
+        # Alternatively print to stdout
+        echo "$1"
     fi
 }
+
+
+########### Pre-flight Checks ##############
+DOCKER_COMPOSE_CONFIG_CHECK=$(docker compose --profile all config 2>&1 > /dev/null)
+if [ $? -ne 0 ]; then
+    ERROR=" :
+\`docker compose --profile all config\`
+> ❌ $DOCKER_COMPOSE_CONFIG_CHECK
+"
+    MESSAGE="${MESSAGE_ERRUPDATE/__DOCKERIMAGE__/$ERROR}"
+    send_telegram_message "$MESSAGE"
+    exit 1
+fi
 
 
 ############ Announcements ###############
@@ -120,10 +136,13 @@ END{
   printf "%.1fGB: ",sum; sep="";
   for(i in n){printf "%s%s",sep,i; sep=", "}
 }')
-MESSAGE="${MESSAGE_CLEANUP/__STATS__/$UNUSED_IMAGES}"
-send_telegram_message "$MESSAGE"
-docker image prune -a -f
-docker builder prune -f
+# Only send message and prune if there are unused images
+if [[ ! "$UNUSED_IMAGES" =~ ^0\.0 ]]; then
+  MESSAGE="${MESSAGE_CLEANUP/__STATS__/$UNUSED_IMAGES}"
+  send_telegram_message "$MESSAGE"
+  docker image prune -a -f
+  docker builder prune -f
+fi
 
 
 ########## Update all Docker images ##########
@@ -164,7 +183,7 @@ for container in $(docker compose ps --services --status=running)
 do
     MESSAGE="${MESSAGE_RESTARTING/__DOCKERSERVICE__/$container}"
     send_telegram_message "$MESSAGE"
-    docker compose up -d --build "$container" 2>&1
+    docker compose up -d --remove-orphans --build "$container" 2>&1
 done
